+++ src/SoftTissue.Application/Controllers/QuasiLinearViscoelasticityController.cs
@@ -2,7 +2,9 @@
 using Microsoft.AspNetCore.Mvc;
 using SoftTissue.Application.Extensions;
 using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress.FungModel;
+using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.GenerateDomain.FungModel;
 using SoftTissue.DataContract.QuasiLinearViscoelasticity.CalculateStress;
+using SoftTissue.DataContract.QuasiLinearViscoelasticity.GenerateDomain;
 using System.Threading.Tasks;
 
 namespace SoftTissue.Application.Controllers
@@ -49,6 +51,19 @@ namespace SoftTissue.Application.Controllers
             [FromBody] CalculateQuasiLinearViscoelasticityStressSensitivityAnalysisExplicitRequest request)
         {
             CalculateQuasiLinearViscoelasticityStressResponse response = await calculateFungModelStressSentivityAnalysis.Process(request.MapToSensitivityAnalysis()).ConfigureAwait(false);
+            return response.BuildHttpResponse();
+        }
+
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status501NotImplemented)]
+        [HttpPost("calculate-stress/fung-model/generate-domain")]
+        public async Task<ActionResult<GenerateDomainResponse>> GenerateFungModelDomain(
+            [FromServices] IGenerateFungModelDomain generateFungModelDomain,
+            [FromBody] GenerateDomainRequest request)
+        {
+            GenerateDomainResponse response = await generateFungModelDomain.Process(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
     }