+++ src/SoftTissue.Application/Startup.cs
@@ -11,6 +11,7 @@ using SoftTissue.Core.NumericalMethods.Integral.Simpson;
 using SoftTissue.Core.Operations.LinearViscoelasticity.CalculateStrain.MaxwellModel;
 using SoftTissue.Core.Operations.LinearViscoelasticity.CalculateStress.MaxwellModel;
 using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress.FungModel;
+using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.GenerateDomain.FungModel;
 
 namespace SoftTissue.Application
 {
@@ -50,6 +51,7 @@ namespace SoftTissue.Application
             services.AddScoped<ICalculateMaxwellModelStressSensitivityAnalysis, CalculateMaxwellModelStressSensitivityAnalysis>();
             services.AddScoped<ICalculateFungModelStress, CalculateFungModelStress>();
             services.AddScoped<ICalculateFungModelStressSentivityAnalysis, CalculateFungModelStressSentivityAnalysis>();
+            services.AddScoped<IGenerateFungModelDomain, GenerateFungModelDomain>();
 
             // Register Numerical Methods
             services.AddTransient<ISimpsonRuleIntegration, SimpsonRuleIntegration>();