+++ src/SoftTissue.Core/Operations/QuasiLinearViscoelasticity/GenerateDomain/FungModel/IGenerateFungModelDomain.cs
@@ -1,0 +1,9 @@
+﻿using SoftTissue.Core.Operations.Base;
+using SoftTissue.DataContract.QuasiLinearViscoelasticity.GenerateDomain;
+
+namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.GenerateDomain.FungModel
+{
+    public interface IGenerateFungModelDomain : IOperationBase<GenerateDomainRequest, GenerateDomainResponse, GenerateDomainResponseData>
+    {
+    }
+}