+++ src/SoftTissue.Core/Operations/QuasiLinearViscoelasticity/GenerateDominium/FungModel/IGenerateFungModelDominium.cs
@@ -1,9 +1,0 @@
-﻿using SoftTissue.Core.Operations.Base;
-using SoftTissue.DataContract.QuasiLinearViscoelasticity.GenerateDominium;
-
-namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.GenerateDominium.FungModel
-{
-    public interface IGenerateFungModelDominium : IOperationBase<GenerateDominiumRequest, GenerateDominiumResponse, GenerateDominiumResponseData>
-    {
-    }
-}