+++ src/SoftTissue.DataContract/QuasiLinearViscoelasticity/GenerateDominium/GenerateDominiumResponseData.cs
@@ -1,8 +1,0 @@
-﻿using SoftTissue.DataContract.OperationBase;
-
-namespace SoftTissue.DataContract.QuasiLinearViscoelasticity.GenerateDominium
-{
-    public class GenerateDominiumResponseData : OperationResponseData
-    {
-    }
-}